Move manager position and fix some test case error.
